IN_PROGRESS/LLAMA_VISION_FORWARD/pyproject.toml
[project]
name = "llama-vision-forward"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.0"
dependencies = [
    "accelerate>=1.6.0",
    "blobfile>=3.0.0",
    "einops>=0.8.1",
    "jax[cuda12]>=0.5.2",
    "matplotlib>=3.10.3",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "regex>=2024.11.6",
    "safetensors>=0.5.3",
    "uv>=0.6.9",
]
